✨ Feat :  기능에 알맞게 네이밍 수정
