+++ src/feature/map/ui/organisms/CurrentLocationSearch.tsx
@@ -1,0 +1,36 @@
+import React from 'react';
+
+import { Pressable, Text, View } from 'react-native';
+
+import ReplayIcons from '@/shared/icons/ReplayIcon';
+import { commonShadow } from '@/styles/shadow';
+
+interface Props {
+  showSearchButton: boolean;
+  onLocationSearch: () => void;
+}
+
+const CurrentLocationSearch = ({
+  showSearchButton,
+  onLocationSearch,
+}: Props) => {
+  return (
+    // ui margin 값 수정 필요합니다
+    // 피그마 코멘트 대응 필요
+    <View>
+      {showSearchButton && (
+        <Pressable
+          className="h-[40px] w-[149px] flex-shrink-0 flex-row items-center justify-center self-center rounded-[27px] bg-neutral-white"
+          onPress={onLocationSearch}
+          style={commonShadow}>
+          <ReplayIcons height={24} width={24} shape="default" />
+          <Text className="ml-1 text-semantic-info body-rg-02">
+            현 지도에서 검색
+          </Text>
+        </Pressable>
+      )}
+    </View>
+  );
+};
+
+export default CurrentLocationSearch;
